requirements.txt
jittor
Pillow
numpy 
matplotlib
yaml
tensorboardX